requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiogram==3.2.0
motor==3.6.0
pymongo==4.9.2
dnspython==2.4.2
pydantic
python-dotenv==1.0.0